pyproject.toml
[tool.poetry]
name = "aws-fastapi-todo-app"
version = "1.0.0"
description = "Advanced REST API with FastAPI on top of API Gateway and Lambda Functions for a TODO app"
authors = ["Santiago Garcia Arango <san99tiago@gmail.com>"]
license = "Apache"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aws-cdk-lib = "2.118.0"
constructs = ">=10.0.0,<11.0.0"

[tool.poetry.group.dev.dependencies]
poethepoet = "^0.24.0"
pytest = "^7.4.4"
pytest-mock = "^3.12.0"
coverage = "^7.4.0"
black = "^23.12.1"
boto3 = "^1.34.14"
moto = "^4.2.13"
aws-lambda-powertools = { version = "^2.31.0" }
python-ulid = "^2.2.0"
jsonschema = "^4.20.0"
fastapi = { extras = ["all"], version = "^0.109.0" }
mangum = "^0.17.0"
pydantic = "^2.5.3"

[tool.pytest.ini_options]
minversion = "7.0"
pythonpath = ["backend", "src"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
black-format = "black ."
test-unit = ["_test_unit", "_coverage_html"]
synth = "cdk synth"
deploy = "cdk deploy --require-approval never"
black-check = "black . --check --diff -v"
_test_unit = "coverage run -m pytest tests/unit"
_coverage_html = "coverage html"

[tool.coverage.run]
branch = true
source = ["backend", "cdk"]
omit = ["**/__init__.py"]

[tool.coverage.report]
show_missing = false
